Created GeocodeService
Created GeocodeService class and added it to app context. The main idea is that we would want to switch out the real geocoder with a mock while testing. This way we can better isolate test and save money on unnecessary api calls.
