+++ src/app.ts
@@ -3,10 +3,12 @@ import bodyParser from 'koa-bodyparser'
 import router from './routes/router'
 import { errorMiddleware } from './middlewares/error_middleware'
 import type { IRepositories } from './respositories'
+import { type IGeocoderService } from './services/geocoderService'
 
 declare module 'koa' {
   interface DefaultContext {
     repositories: IRepositories
+    geocodeService: IGeocoderService
   }
 }
