+++ src/server.ts
@@ -2,10 +2,12 @@ import app from './app'
 import env from './lib/env'
 import { connectToDatabase, disconnectDatabase } from './database'
 import Repositories from './respositories'
+import { buildGeocodeService } from './services/geocoderService'
 
 const startServer = async (): Promise<void> => {
   try {
     await connectToDatabase()
+    app.context.geocodeService = buildGeocodeService()
     app.context.repositories = Repositories
     app.on('error', (err, ctx) => {
       console.error('Server error', err, ctx)
